renames component and moves to routes
